feat: Replace react-input-mask with react-imask
The react-input-mask package was causing erratic behavior and a poor
user experience when typing in masked input fields. This packege will be
needed to create a good experience for the user data editing feature.

This commit migrates the input masking functionality to react-imask, a
more modern and actively maintained library that provides a smoother and
more reliable typing experience.

The following changes were made:
 - Removed react-input-mask from dependencies.
 - Added react-imask to dependencies.

Updated the input components on the Contact and QuoteRequest pages to
use the new IMaskInput component.